- Culling any files that were at the repo level in azure-service-bus-go but are no longer needed in the monorepo - Changing any package references and the go.mod file to reflect the new location
