--- sdk/messaging/internal/amqp-common/azure-pipelines.yml
@@ -1,95 +1,0 @@
-variables:
-  GOPATH: '$(system.defaultWorkingDirectory)/work'
-  sdkPath: '$(GOPATH)/src/github.com/$(build.repository.name)'
-  GO111MODULE: 'on'
-
-jobs:
-  - job: 'azureamqpcommongo'
-    displayName: 'Run azure-amqp-common-go CI Checks'
-
-    strategy:
-      matrix:
-        Linux_Go113:
-          vm.image: 'ubuntu-18.04'
-          go.version: '1.14.10'
-        Linux_Go114:
-          vm.image: 'ubuntu-18.04'
-          go.version: '1.15.3'
-
-    pool:
-      vmImage: '$(vm.image)'
-
-    steps:
-      - task: GoTool@0
-        inputs:
-          version: '$(go.version)'
-        displayName: "Select Go Version"
-
-      - script: |
-          set -e
-          mkdir -p '$(GOPATH)/bin'
-          mkdir -p '$(sdkPath)'
-          shopt -s extglob
-          mv !(work) '$(sdkPath)'
-          echo '##vso[task.prependpath]$(GOPATH)/bin'
-          go version
-        displayName: 'Create Go Workspace'
-
-      - script: |
-          set -e
-          go get github.com/jstemmer/go-junit-report
-          go get github.com/axw/gocov/gocov
-          go get github.com/AlekSi/gocov-xml
-          go get -u github.com/matm/gocov-html
-          go get -u golang.org/x/lint/golint
-          go get github.com/fzipp/gocyclo/cmd/gocyclo
-        workingDirectory: '$(sdkPath)'
-        displayName: 'Install Dependencies'
-
-      - script: |
-          go build -v ./...
-        workingDirectory: '$(sdkPath)'
-        displayName: 'Build'
-
-      - script: |
-          go vet ./...
-        workingDirectory: '$(sdkPath)'
-        displayName: 'Vet'
-
-      - script: |
-          golint ./...
-        workingDirectory: '$(sdkPath)'
-        displayName: 'Lint'
-
-      - script: |
-          gocyclo -over 19 .
-        workingDirectory: '$(sdkPath)'
-        displayName: 'Cyclo'
-
-      - script: |
-          set -e
-          go test -race -v -coverprofile=coverage.txt -covermode atomic ./... 2>&1 | go-junit-report > report.xml
-          gocov convert coverage.txt > coverage.json
-          gocov-xml < coverage.json > coverage.xml
-          gocov-html < coverage.json > coverage.html
-        workingDirectory: '$(sdkPath)'
-        displayName: 'Run Tests'
-
-      - script: |
-          gofmt -s -l -w . >&2
-        workingDirectory: '$(sdkPath)'
-        displayName: 'Format Check'
-        failOnStderr: true
-        condition: succeededOrFailed()
-
-      - task: PublishTestResults@2
-        inputs:
-          testRunner: JUnit
-          testResultsFiles: $(sdkPath)/report.xml
-          failTaskOnFailedTests: true
-
-      - task: PublishCodeCoverageResults@1
-        inputs:
-          codeCoverageTool: Cobertura 
-          summaryFileLocation: $(sdkPath)/coverage.xml
-          additionalCodeCoverageFiles: $(sdkPath)/coverage.html
