--- sdk/messaging/internal/amqp-common/internal/version.go
@@ -1,6 +1,0 @@
-package common
-
-const (
-	// Version is the semantic version of the library
-	Version = "3.1.1"
-)
