+++ sdk/messaging/internal/auth/token.go
