+++ sdk/messaging/internal/go.mod
@@ -1,4 +1,4 @@
-module github.com/Azure/azure-amqp-common-go/v3
+module github.com/Azure/azure-sdk-for-go/sdk/messaging/internal
 
 go 1.12
