+++ sdk/messaging/internal/ptrs.go
@@ -1,4 +1,6 @@
-package common
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+package internal
 
 //	MIT License
 //
