+++ sdk/messaging/internal/version.go
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+package internal
+
+const (
+	// Version is the semantic version of the library
+	Version = "3.1.1"
+)
